# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API REST - OAuth2 & GIPHY Integration",
        "description": "API REST construida con Laravel 12, OAuth2.0 (Passport), Doctrine ORM y Redis Cache, siguiendo arquitectura hexagonal.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/media/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Obtener GIF por ID (con Redis Cache)",
                "description": "Obtiene un GIF específico de GIPHY por su ID. Las respuestas se cachean en Redis (Cache key: media:id:{id}).",
                "operationId": "3aa47443d80ba01b5944c6dd8cf8e665",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del GIF en GIPHY",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123xyz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GIF encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Media encontrado exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "abc123xyz"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Funny Cat GIF"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://giphy.com/gifs/abc123xyz"
                                                },
                                                "rating": {
                                                    "type": "string",
                                                    "example": "g"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "catlovers"
                                                },
                                                "images": {
                                                    "properties": {
                                                        "original": {
                                                            "properties": {
                                                                "url": {
                                                                    "type": "string",
                                                                    "example": "https://media.giphy.com/media/abc123/giphy.gif"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "preview_gif": {
                                                            "properties": {
                                                                "url": {
                                                                    "type": "string",
                                                                    "example": "https://media.giphy.com/media/abc123/200.gif"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "GIF no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Media no encontrado"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Media with ID \"abc123xyz\" not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Error al conectar con GIPHY API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al conectar con el proveedor de media"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT en formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}